+++ src/main/kotlin/icu/windea/pls/PlsIcons.kt
@@ -26,6 +26,7 @@ val localisationCommandFieldIcon = IconLoader.getIcon("/icons/localisationComman
 val definitionIcon = IconLoader.getIcon("/icons/definition.svg",locationClass)
 val definitionLocalisationIcon = IconLoader.getIcon("/icons/definitionLocalisation.svg",locationClass)
 val localisationIcon = IconLoader.getIcon("/icons/localisation.svg",locationClass)
+val propertyIcon = IconLoader.getIcon("/icons/property.svg", locationClass)
 val valueIcon = IconLoader.getIcon("/icons/value.svg",locationClass)
 val enumIcon = IconLoader.getIcon("/icons/enum.svg",locationClass)
 val aliasIcon = IconLoader.getIcon("/icons/alias.svg",locationClass) //目前用不到，因为名字是表达式