+++ src/main/kotlin/icu/windea/pls/localisation/psi/ParadoxLocalisationNameIndex.kt
@@ -152,7 +152,7 @@ object ParadoxLocalisationNameIndex : StringStubIndexExtension<ParadoxLocalisati
 	}
 	
 	private fun matches(key: String, keyword: String): Boolean {
-		return key.matchesKeyword(keyword,true)
+		return key.matchesKeyword(keyword)
 	}
 }